refactor: create extra file for messageHandler
